refactor to hook
